fixed remeasure uncertainty
before, due to constant conversions in image formats, the images changes ever-so-slightly between each new set of measurements, meaning if you asked to remeasure the area, it would do so on a slightly different image, giving subtly different measurements. This was fixed by simply not changing the "originalImage"- before it was changed after each analysis which actually was not necessary.